e2e: Do not run TestShutdown on Windows
